CronJob, EndpointSlice and PodDisruptionBudget moved to v1
