+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/dsl/base/OperationContextTest.java
@@ -16,8 +16,8 @@
 package io.fabric8.kubernetes.client.dsl.base;
 
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
-import io.fabric8.kubernetes.api.model.batch.Job;
-import io.fabric8.kubernetes.api.model.batch.JobBuilder;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
+import io.fabric8.kubernetes.api.model.batch.v1.JobBuilder;
 import io.fabric8.kubernetes.client.Config;
 import okhttp3.OkHttpClient;
 import org.junit.jupiter.api.BeforeEach;
