+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/dsl/internal/batch/v1/JobOperationsImplTest.java
@@ -15,8 +15,8 @@
  */
 package io.fabric8.kubernetes.client.dsl.internal.batch.v1;
 
-import io.fabric8.kubernetes.api.model.batch.Job;
-import io.fabric8.kubernetes.api.model.batch.JobBuilder;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
+import io.fabric8.kubernetes.api.model.batch.v1.JobBuilder;
 import org.junit.jupiter.api.Test;
 
 import java.util.Map;
