+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/JobIT.java
@@ -15,8 +15,8 @@
  */
 package io.fabric8.kubernetes;
 
-import io.fabric8.kubernetes.api.model.batch.Job;
-import io.fabric8.kubernetes.api.model.batch.JobBuilder;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
+import io.fabric8.kubernetes.api.model.batch.v1.JobBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.LogWatch;
 import org.arquillian.cube.kubernetes.api.Session;
