+++ kubernetes-model-generator/kubernetes-model-batch/cmd/generate/generate.go
@@ -61,6 +61,8 @@ type Schema struct {
 
   Job                                      batchapiv1.Job
   JobList                                  batchapiv1.JobList
+  V1CronJob                                batchapiv1.CronJob
+  V1CronJobList                            batchapiv1.CronJobList
   CronJob                                  batchapiv1beta1.CronJob
   CronJobList                              batchapiv1beta1.CronJobList
 }
@@ -72,8 +74,8 @@ func main() {
     {"k8s.io/apimachinery/pkg/version", "", "io.fabric8.kubernetes.api.model.version", "kubernetes_apimachinery_pkg_version_", false},
     {"k8s.io/apimachinery/pkg/apis/meta/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_apimachinery_", false},
     {"k8s.io/api/core/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_core_", false},
-    {"k8s.io/api/batch/v1beta1", "", "io.fabric8.kubernetes.api.model.batch", "kubernetes_batch_", true},
-    {"k8s.io/api/batch/v1", "", "io.fabric8.kubernetes.api.model.batch", "kubernetes_batch_", true},
+    {"k8s.io/api/batch/v1beta1", "", "io.fabric8.kubernetes.api.model.batch.v1beta1", "kubernetes_batch_v1beta1_", true},
+    {"k8s.io/api/batch/v1", "", "io.fabric8.kubernetes.api.model.batch.v1", "kubernetes_batch_v1_", true},
   }
 
   typeMap := map[reflect.Type]reflect.Type{
